Add containers, components, and utils folders for Main and move constants to separate file
